backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.3
python-docx==1.1.2
google-generativeai==0.7.2
python-dotenv==1.0.1
scipy==1.13.1
scikit-learn==1.5.0
Werkzeug==3.0.3
gunicorn==22.0.0
matplotlib==3.9.0
seaborn==0.13.2
